Refactor internal util.SysReadFile()
* Move SysReadFile() function to a non-windows build file.
* Add build hint for non-windows to sysfs functions.
